MOD: mv __init__.py to __main__.py
